test(desktop): share the form-control helpers and blur before the git probe
Three more harness defects from the second run, all reporting the product as
broken over a screenshot of it working.

`selectOption` read its options as WDIO handles: `$$` resolves to a chainable
array-like whose `map` is not the array method, so awaiting the mapped result
threw "object is not iterable" from inside the helper. The values are now read
in the page in one round trip.

`fill` existed twice. The counted-Backspace version fixed in the settings spec
was still live in the workspace spec, so the same off-by-one was one edit away
from returning. Both specs now import one implementation.

The create-session dialog inspects a typed path on blur, and a bare `Tab` did
not move focus off the field -- the failure screenshot shows it still ringed
with neither the "Git project" line nor an error under it, so `inspectPath` had
never run. The spec now clicks the dialog heading, and waits for the inspection
result to render before reading the worktree checkbox, so "has not looked yet"
stops being indistinguishable from "looked and said no".
