+++ tests/desktop/helpers/form-control.mjs
@@ -1,4 +1,35 @@
 import assert from "node:assert/strict";
+import { Key } from "webdriverio";
+
+/**
+ * Types `value` into a controlled input, clearing it with the keyboard first.
+ *
+ * Not `setValue`: that issues WebDriver's Element Clear, which assigns `value` directly. React's
+ * controlled-input value tracker can swallow the event that follows, leaving component state on
+ * the old text while the box on screen looks empty -- the submit then sends the stale value and
+ * the test "passes" against a field it never actually changed.
+ *
+ * Not one Backspace per character either. A counted run is off by one whenever a keystroke lands
+ * before the caret settles from the click, and the residue rides into the typed value: clearing a
+ * 12-character field with 12 Backspaces produced "Ivanehub-ui-settings-e2e", which reads as the
+ * app mangling input rather than as the test miscounting. Ctrl+A does not depend on the caret
+ * position or on how many keys arrive.
+ */
+export async function fill(description, element, value) {
+  await element.waitForClickable({ timeout: 15_000 });
+  await element.click();
+  const current = String((await element.getProperty("value")) ?? "");
+  if (current.length > 0) {
+    await globalThis.browser.keys([Key.Ctrl, "a"]);
+    await globalThis.browser.keys(["Backspace"]);
+  }
+  if (value.length > 0) await element.addValue(value);
+  await globalThis.browser.waitUntil(
+    async () => (await element.getProperty("value")) === value,
+    { timeout: 15_000, timeoutMsg: `${description} did not accept the typed value.` },
+  );
+}
+
 
 /**
  * Moves a native `<select>` to `value` with the keyboard and waits for the app to hold it.
@@ -16,8 +47,14 @@ import assert from "node:assert/strict";
  * that arrives during one is dropped.
  */
 export async function selectOption(description, element, value) {
-  const options = await element.$$("option");
-  const values = await Promise.all(options.map((option) => option.getAttribute("value")));
+  // Read the option values in the page rather than as WDIO element handles: `$$` resolves to a
+  // chainable array-like whose `map` is not the array method, so mapping it and awaiting the
+  // result throws "object is not iterable" from inside this helper -- an error about the helper,
+  // reported against the widget.
+  const values = await globalThis.browser.execute(
+    (select) => Array.from(select.options).map((option) => option.value),
+    element,
+  );
   const target = values.indexOf(value);
   assert.ok(target >= 0, `${description} has no ${value} option; it offers ${values.join(", ")}`);
